requirements.txt
attrs==22.2.0
click==8.1.3
colorama==0.4.6
DateTime==5.1
Flask==2.2.3
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.7
threadpoolctl==3.1.0
typing_extensions==4.5.0
uuid==1.30
Werkzeug==2.2.3
zope.interface==6.0
